detection sans erreur

+++ src/detection/train/EtatProbable.java
@@ -2,8 +2,8 @@ package detection.train;
 
 public class EtatProbable {
 		public int Idetat;
-		public int proba;
-		public EtatProbable(int idetat, int proba) {
+		public double proba;
+		public EtatProbable(int idetat, double proba) {
 			super();
 			Idetat = idetat;
 			this.proba = proba;
